Add PHP pdo_pgsql

+++ include/pecl_pgsql.sh
@@ -18,8 +18,14 @@ Install_pecl-pgsql() {
   ./configure --with-pgsql=${pgsql_install_dir} --with-php-config=${php_install_dir}/bin/php-config
   make -j ${THREAD} && make install
   popd
-  if [ -f "${phpExtensionDir}/pgsql.so" ]; then
+  pushd php-${PHP_detail_version}/ext/pdo_pgsql
+  ${php_install_dir}/bin/phpize
+  ./configure --with-pdo-pgsql=${pgsql_install_dir} --with-php-config=${php_install_dir}/bin/php-config
+  make -j ${THREAD} && make install
+  popd
+  if [ -f "${phpExtensionDir}/pgsql.so" -a -f "${phpExtensionDir}/pdo_pgsql.so" ]; then
     echo 'extension=pgsql.so' > ${php_install_dir}/etc/php.d/ext-pgsql.ini
+    echo 'extension=pdo_pgsql.so' >> ${php_install_dir}/etc/php.d/ext-pgsql.ini
     echo "${CSUCCESS}PHP pgsql module installed successfully! ${CEND}"
     popd
     rm -rf php-${PHP_detail_version}
